Add single region US db and template file
Includes most recent version of the singele-region US database with 12 time-slices.
This database is copied over from the data repo.
Included here is a template file that can be used to create a new US database for the OEO.
The "commodities" table contains a list of commodities from the current version of the single-region US database.
All the other tables are empty, and can be used to enter data for the different OEO sectors.